# crates/cosmon-rpp-adapter/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Cosmon Remote Pilot Port (RPP) — V0+V1
  version: "1.0.0"
  description: |
    Hand-curated tenant client contract for the Remote Pilot Port. Requests
    execute through the admission boundary and injectable library ports.
    Worker dispatch may spawn processes; tenant state remains isolated.

    The registered surface is partitioned between this document and exact
    entries in openapi/exclusions.json, each with a reviewed reason. The
    cargo integration test openapi_surface checks both directions of path
    coverage directly against src/lib.rs, independently of the event canon.
    This file is not generated from Rust types. Path coverage does not prove
    method, schema or behavioral parity; handler changes still require review.

    Bounded run (ADR-124) and sealed done (ADR-176 as amended) are tenant
    request doors. Their presence does not expose the whole CLI. Operator
    provisioning, bootstrap distribution and the separate MCP protocol are
    outside this published REST contract. Existing healthz, metrics and
    diagnostics remain documented operational endpoints without tenant data.
  contact:
    name: Noogram
    url: https://github.com/noogram/cosmon
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
servers:
  - url: https://rpp.{domain}
    description: Operator-controlled deployment (typical tenant_auditor-facing instance).
    variables:
      domain:
...
security:
  - bearerAuth: []
paths:
...
  /v1/events:
    get:
      operationId: streamEvents
      summary: |
        Subscribe to the per-tenant Server-Sent Events stream of
        molecule lifecycle events (`task-20260522-c46a`).
      description: |
        Long-lived `text/event-stream` response. The adapter publishes
        a `MoleculeEvent` to an in-process bus from every mutation
        route (`nucleate`, `tag`, `collapse`, `freeze`, `stuck`,
        `tackle`); the SSE handler subscribes to the bus and forwards
        events matching the caller's tenant `noyau` (and optionally a
        single `molecule_id`) to the wire.

        ## Required scope

        `cosmon:events:subscribe` (distinct from
        `cosmon:molecule:read` — a long-lived SSE tail is a different
        blast radius than periodic observe).

        ## Wire format

        Each event arrives as:

        ```
        id: <monotonic-u64>
        event: <name>
        data: <json>

        ```

        Two event names are emitted today (the set is additive — a
        v1 subscriber MUST ignore unknown names rather than reject):

        - `molecule.state_changed` —
          `{"molecule_id": "...", "old_state": "...", "new_state": "...", "timestamp": "..."}`
        - `molecule.event_appended` —
          `{"molecule_id": "...", "event": {...}, "timestamp": "..."}`

        A keep-alive comment line (`: keep-alive`) is sent every 30 s
        so any HTTP/1.1 proxy in the middle keeps the socket open.

        ## Reconnection

        The SSE channel is a *live tail*: history is **not** replayed
        server-side. A client that drops and reconnects with
        `Last-Event-ID: <n>` will have events with `id <= n` filtered
        out at the adapter, but events that landed on the bus between
        disconnect and reconnect are gone. Durable history lives in
        the per-tenant filesystem state.

        ## Cross-tenant isolation

        Structural — every event carries an internal `noyau` field
        (skipped on serialisation); the handler subscribes after
        resolving the JWT's nucleon and refuses to forward events
        whose `noyau` does not match.
      tags: [events]
      parameters:
        - in: query
          name: molecule_id
          required: false
          schema:
            type: string
          description: |
            Filter to a single molecule id. When omitted, every event
            for the caller's noyau is streamed.
        - in: header
          name: Last-Event-ID
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
          description: |
            Resume floor — events with `id <= Last-Event-ID` are
            filtered out before they reach the wire. No replay.
      responses:
        "200":
          description: |
            SSE stream — the connection stays open until the client
            disconnects, the server is shut down, or the network
            drops.
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE-encoded event stream (see description).
        "400":
          description: Malformed query parameter (framework rejection).
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Missing or invalid JWT or tenant admission failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Missing cosmon:events:subscribe scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: rate_limited or noyau_budget_exhausted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/ledger:
    get:
      operationId: streamLedger
      summary: |
        Replay, then live, over the tenant's durable event log, resumable
        from an opaque cursor (issue #184).
      description: |
        Long-lived `text/event-stream` response over the tenant's ledger.
        One reader reads the log to its end and then keeps following the
        same position, so replay and live share one position: no line falls
        between them and none is delivered twice. Lines appended by workers
        are in the ledger and therefore on this stream, unlike
        `GET /v1/events`, which only carries what the adapter's own
        handlers publish and replays nothing.

        ## Required scope

        Both `cosmon:events:subscribe` and `cosmon:molecule:read`. The
        second is required because `molecule_step_completed.evidence` is
        free text a worker wrote, which no other event stream serves.

        ## Cursor

        The SSE `id` of every data frame is the resume cursor: an opaque
        token, to be echoed back verbatim. It is not the envelope `seq`,
        which lifecycle lines written by the file store do not carry. Resume
        with `Last-Event-ID` (what a reconnecting `EventSource` sends) or
        `after`; the header wins when both are present. With neither, the
        stream starts at the beginning of the ledger.

        ## Frames

        - `ledger.epoch` — always first. `{"epoch", "cursor"}` names the
          log the cursors of this connection refer to.
        - `ledger.reset` — the consumer's folded state is stale and the
          stream replays from the start of the log named by `cursor`.
          `{"reason", "epoch", "cursor"}`; `reason` is `epoch_changed`
          (another file: rotation or replacement, also sent when a
          supplied cursor names a different file), `truncated` (the file
          shrank) or `cursor_invalid` (malformed, past the end, or not on
          a line boundary). Never silent.
        - one frame per served ledger line, `event` = the line's type and
          `data` a projection of its documented fields:
          `schema_version` (the line's own, `1` when the line carries
          none), `type`, `molecule_id`, `timestamp`, `seq` and `mol_seq`
          when the line has them (information only), and the fields listed
          for the type in `docs/book/src/reference/read-contracts.md`.
          Types outside that table, lines naming no molecule
          (`operator_present`, `operator_signed`), worktree paths, shell
          commands and harness-turn text are never served; `session_id`
          is replaced by `session_digest`. The set of types is additive:
          ignore unknown event names.

        ## Bounds

        A connection delivers at most 5,000 frames, then closes cleanly;
        reconnect with the last `id`. At most one ledger stream per
        principal is open at a time (a second is refused with 429). One
        rate-limit token is charged per connection. Admission is rechecked
        every second, as on the other streams.
      tags: [events]
      parameters:
        - in: query
          name: after
          required: false
          schema:
            type: string
          description: |
            Resume cursor: the `id` of the last frame received. Overridden
            by `Last-Event-ID` when both are sent.
        - in: header
          name: Last-Event-ID
          required: false
          schema:
            type: string
          description: Resume cursor, verbatim from the last frame's `id`.
      responses:
        "200":
          description: |
            SSE stream; closes after at most 5,000 frames, or when the
            client disconnects or admission ends.
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE-encoded frames (see description).
                x-sse-data-schema:
                  oneOf:
                    - $ref: "#/components/schemas/LedgerEpochFrame"
                    - $ref: "#/components/schemas/LedgerResetFrame"
                    - $ref: "#/components/schemas/LedgerEventFrame"
        "401":
          description: Missing or invalid JWT or tenant admission failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Missing `cosmon:events:subscribe` or `cosmon:molecule:read`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: The caller's tenant has no galaxy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: |
            `rate_limited`, `noyau_budget_exhausted`, or
            `ledger_stream_open` (this principal already holds a stream).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The ledger cannot be read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  # ── Operational endpoints (outside /v1/, no JWT gate, excluded
  # from the §8p frozen surface). Documented here for completeness so
  # a dashboard tenant can render the adapter health page without
  # out-of-band discovery (task-20260523-d820, gap report ae3d §i).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC-issued JWT signed with `RS256` or `ES256` (clause (a)).
        The token's `kid` MUST resolve in the operator's pinned JWKS
        (`<state_dir>/security/jwks/`). `none`, `HS256`, `EdDSA`,
        `RS384`, `RS512` are rejected at parse time.
  schemas:
...
    LedgerEpochFrame:
      type: object
      additionalProperties: true
      required: [epoch, cursor]
      properties:
        epoch: { type: string }
        cursor: { type: string }
    LedgerResetFrame:
      type: object
      additionalProperties: true
      required: [reason, epoch, cursor]
      properties:
        reason:
          type: string
          enum: [epoch_changed, truncated, cursor_invalid]
        epoch: { type: string }
        cursor: { type: string }
    LedgerEventFrame:
      type: object
      additionalProperties: true
      required: [schema_version, type, molecule_id]
      properties:
        schema_version: { type: integer, minimum: 1 }
        type: { type: string }
        molecule_id: { type: string }
        timestamp: { type: string, format: date-time }
        seq: { type: integer, minimum: 0 }
        mol_seq: { type: integer, minimum: 0 }
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: |
            Coarse-grained error label.

            **Open by contract.** The server may emit new labels in
            future versions without an API bump. Callers MUST treat
            any unknown value as semantically equivalent to
            `internal` (i.e. retry-on-idempotent, surface-as-opaque).
            Known values at v1.0.0 include `unauthorized`,
            `forbidden`, `not_found`, `invalid_filter`,
            `invalid_json_body`, `missing_reason`, `empty_reason`,
            `empty_tag_request`, `terminal_status`, `invalid_status`,
            `rate_limited`, `store_unavailable`, `endpoint_removed`,
            `internal` — but this list is *informative*, not
            normative. Mirror of the lenient `#[non_exhaustive]` +
            `#[serde(other)] Unknown(String)` discipline already
            inscribed for `Status` / `Kind` / `MoleculeEventKind`.

            The wire body never carries `sub`, `nucleon_id`, or any
            tenant-identifying field.
        request_id:
          type: string
          description: |
            Optional correlation id; present on errors that
            materialise an audit envelope.
        retry_after_seconds:
          type: integer
          format: int64
          minimum: 0
          description: |
            Present only when `error` is `rate_limited`. The number of
            whole seconds to wait before retrying, rounded up from the
            admission boundary's precise delay. The `Retry-After` response
            header carries this same decimal value; clients may use either
            representation and MUST NOT retry a non-idempotent mutation
            without an idempotency key.
...